Create tex file from executive summary tables

Create tex file from executive summary tables

es_table_tex(
  dir,
  table_folder = NULL,
  save_loc = NULL,
  csv_name = "table_labels.csv"
)

Arguments

dir

The directory to look for the csv file that has a list of all csv files to create tex files for. The csv file format should match the table_labels.csv created by r4ss::SSexecutivesummary(). The columns names in this folder are caption, altcaption, label, filename, and loc (optional). The loc column is a directory location to find the file (filename) that a tex file should be created for. The dir can be set to the mod_loc where the function will automatically look for them inside a table folder in that location.

table_folder

The relative path for the table directory, where it must be relative to dir because it will be appended to this argument using file.path(). An alternative is to supply the full file path in dir and use table_folder = "".

save_loc

optional input that requires a full path which will allow users to save the tex file to a specific location (e.g. inside doc folder) other than the file.path(dir, table_folder). Default is NULL.

csv_name

CSV file name to create tex table scripts for. This file should have the following columns: caption, altcaption, label, filename, loc (optional) where caption will be the table caption, altcaption is the accessibility text for the table (can be NA if the caption should be used), label is the text to reference the table, filename is the file name of the csv file to be read, and loc (optional) is the location to file the csv file indicated by the filename.

Author

Chantel R. Wetzel

Examples

if (FALSE) { # \dontrun{
# example of creating tex files for those created by the
# r4ss::SSexecutivesummary()
es_table_tex(dir = mod_loc)
# example of using a user created csv file:
es_table_tex(
  dir = "C:/models/table_to_add_to_doc",
  save_loc = "C:/model/tex_tables",
  csv_name = "non_es_document_tables.csv"
)
} # }