sa4ss: Write Stock Assessments for Stock Synthesis

Create files and the structure necessary to facilitate rapid and reproducible creation of stock assessment documents from model output generated by Stock Synthesis (SS3). The included components and structure are specific to requests from the Pacific Fishery Management Council for U.S. west coast groundfish.

See also

Author

Maintainer: Kelli F. Johnson kelli.johnson@noaa.gov (ORCID)

Authors:

Chantel R. Wetzel chantel.wetzel@noaa.gov (ORCID)