Technical report output formatting

Format for creating a technical report with the .pdf extension. This format will be used to generate markdown from R Markdown and a resulting pdf will be saved to the disk.

Usage

techreport_pdf(
  latex_engine = c("lualatex", "pdflatex"),
  pandoc_args = c("--top-level-division=section", "--wrap=none",
    "--default-image-extension=png"),
  ...
)

Arguments

latex_engine

The desired latex engine, where luaLaTex is the default because it leads to a 508 compliant document. This argument is passed to pdf_book. Only a single value can be passed, but all options are listed below.

pandoc_args

A vector of strings that lead to arguments for pandoc. This argument is passed to bookdown::pdf_book(). An optional argument that users may wish to add to the default arguments is "--lua-filter=tagged-filter.lua" which will lead to the package-supplied filter being used to generate tags. This is no longer the preferred method for creating accessible documents as the kernel for LaTeX now includes much of this functionality by default.

...

Additional arguments passed to pdf_book.