Compile Title in YAML format

Write a file to the disk with the title using the yaml form for inclusion in the stock assessment.

write_title(
  species,
  latin,
  coast,
  year = format(Sys.Date(), "%Y"),
  fileout = "00title.Rmd"
)

Arguments

species

A vector of character strings providing species names for all species included in the stock assessment. Typically, there will only be one species, but this template can accommodate a species complex as well. Each species should be capitalized as you would like to see it used in the middle of a sentence. For example, "Pacific Hake" is used in the stock assessment of hake; not "pacific hake" because Pacific Hake is the preferred name, not pacific hake. The text can be called in the document using spp for the form to be used in the middle of a sentence and Spp when placed at the beginning of a sentence.

latin

A character string specifying the Latin name without italic formatting. Please capitalize as you want it to be used in the document though because case will not be altered.

coast

A character string specifying which part of the coast the stock assessment includes. For example, "U.S. West Coast", which is the default, will be used to create the following title: "Status of common name (latin name) along the U.S. West Coast in 2010". Previously, the word coast was added internally to the text, this is no longer the default behavior of sa4ss and instead you must explicitly add the word, which allows users to have more flexibility. For example, "California Coast from Point Conception north" is a viable option.

year

A numeric or character string specifying the year. The default uses Sys.Date() to format the year into a four digit character value.

fileout

A file path for the file that will be saved to the disk. Typically, full or relative paths are allowed.

Details

The Latin name as specified in "latin" argument will be encased in underscores to facilitate it being italicized. Using emph is not allowed because the tagging structure for the creation of the 508 compliant pdf will not render the LaTex code here. Nevertheless, the use of markdown inside the yaml seems to work just fine.

See also

Author

Kelli F. Johnson